clean-up(cache): remove nodes tracking from cache
The original idea was to track to how many nodes a given resource
has been sent in order to know when was the right moment to delete it.
Think about deployments that could have multiple pods to the same node.
This logic is not needed anymore since we will sent a delete event to a
node only when the resource is not related to it in anyway. For example,
a deploymen is deleted for a node when it has no more pods running a that node.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>