fix(dispatch): avoid deadlock at exit time
When a sigterm is received the active goroutines exits, and the subscriber
connections are closed. When a subscriber connection is closed it sends
a message to the dispatcher goroutine saying that it is done. But if the
dispatcher goroutine does not retrieve the messages from the buffered channel
of length 1, causes the subscribers to hang. This fix ensures that the dispatcher
goroutines do not exit without waiting for all the subscriber's connections to be closed.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>